# Fix incorrect keyboard input processing (#942)