Rewrite KEYMAP:LOOKUP-KEY using dolist/return.
Fixes #942