@Ambrevar 2020-09-09
I can't reproduce the issue in #942.  Can you explain why this fixes it?