@dstatyvka 2020-09-09
I'll do my best.

As I described in the issue the original code looses the first return value.  However if I put `(format *trace-output* "result is ~a~%" result)`, everything works just fine I can see "result is ..." in the trace log.  The bug comes back when I change the code back.
Being re-written as suggested it works.

The only explanation think I can think of is some SBCL bug, which I'm unable to reproduce on any simpler example.  At the moment I cannot check it on the last SBCL either.  I think in order to reproduce the bug one has to have an very close environment to mine.