@Ambrevar 2020-09-09
My hunch is that this has to do with a stack corruption because
lookup-key* is called so many times (I'm abusing recursion here).

Home much virtual memory do you have?

If you build Nyxt by passing 

--control-stack-size 2048

to the SBCL flags, does it help?

I'll work on a fix, the lookup-key* calls is what makes
`execute-command` pop up slow anyways.
Maybe I can inline the function or block-compile it.