@dstatyvka 2020-09-09
> Home much virtual memory do you have?
```
$ free -m
              total        used        free      shared  buff/cache   available
Mem:          15904        6609        2173         915        7120        8073
Swap:         32499           0       32499
```
>  If you build Nyxt by passing --control-stack-size 2048 to the SBCL flags, does it help?
Nope, it doesn't.  Checked like this:
```
$ sbcl --control-stack-size 2048
* (ql:quickload :nyxt/gtk-application)
* (nyxt:start)
```
Trying different sizes gives nothing.