@dstatyvka 2020-09-09
@Ambrevar TBH, I don't think there's anything wrong with recursion.  As you can see from the trace output I've listed in #943  there's not that much calls of either `lookup-key*` or `lookup-key`.