@Ambrevar 2020-09-09
> Sometimes I get into a mode with keys where I can only type the letter "S" for some reason. The system stays responsive otherwise. I may try this patch and see if it fixes this intermittent problem.

Weird, never had this problem.

This is getting more and more puzzling.