@dstatyvka 2020-09-10
I cannot reproduce the issue using SBCL 2.0.8