@jmercouris 2020-09-18
Update: I've figured out what my issue is. If I press "meta" and then use my hotkey to switch applications, the state of the "meta" toggle get's messed up and Nyxt believes that "meta" is still pressed since it does not get a key up invocation. This is because of the concessions we have to make on macOS with regards to "meta" being a second class citizen.

@dstatyvka can you please confirm that this patch is not necessary with a recent SBCL?