@jmercouris 2020-09-18
Alternatively, if you can provide a (time ...) comparison, we can take the faster implementation (whether it is master, or this version).