@dstatyvka 2020-09-19
> @dstatyvka can you please confirm that this patch is not necessary with a recent SBCL?

Right, with SBCL 2.0.8 it works for me without the patch