@dstatyvka 2020-09-19
> Alternatively, if you can provide a (time ...) comparison, we can take the faster implementation (whether it is master, or this version).

Okay, I'll give it a shot in a day or two, let's see if I can measure any difference.