@dstatyvka 2020-09-24
This is what I've got using trivial-benchmark
```common-lisp
(defparameter *very-long-list* (loop :repeat 100000 :append (current-keymaps)))
(ql:quickload :trivial-benchmark)
(trivial-benchmark:with-timing (100)
  (keymap:lookup-key "C-x h" *very-long-list*))
```

```
master (c7e444b8)

-                SAMPLES  TOTAL       MINIMUM   MAXIMUM   MEDIAN    AVERAGE     DEVIATION  
REAL-TIME        100      103.869     1.016     1.204     1.031     1.03869     0.029101   
RUN-TIME         100      103.898     1.016     1.205     1.031     1.03898     0.02891    
USER-RUN-TIME    100      103.89337   1.01613   1.204771  1.031557  1.038934    0.028891   
SYSTEM-RUN-TIME  100      0.002863    0         0.002863  0         0.000029    0.000285   
PAGE-FAULTS      100      0           0         0         0         0           0.0        
GC-RUN-TIME      100      0.263       0         0.007     0         0.00263     0.002982   
BYTES-CONSED     100      2401102560  23985936  24051696  24018928  24011026.0  14742.94   
EVAL-CALLS       100      0           0         0         0         0           0.0        
```

```
patched (af9d1a95)
-                SAMPLES  TOTAL       MINIMUM   MAXIMUM   MEDIAN    AVERAGE     DEVIATION  
REAL-TIME        100      104.29      1.025     1.106     1.04      1.0429      0.012923   
RUN-TIME         100      104.308     1.024     1.107     1.04      1.04308     0.01299    
USER-RUN-TIME    100      104.29588   1.024577  1.107027  1.040245  1.042959    0.013044   
SYSTEM-RUN-TIME  100      0.011383    0         0.011383  0         0.000114    0.001133   
PAGE-FAULTS      100      0           0         0         0         0           0.0        
GC-RUN-TIME      100      0.255       0         0.007     0         0.00255     0.002903   
BYTES-CONSED     100      2401135712  23985952  24051696  24018928  24011358.0  14551.46   
EVAL-CALLS       100      0           0         0         0         0           0.0        
```

I would say there's no significant difference in performance